// Frontend/src/generated/watchlists/model/watchlist.ts
/**
 * Watchlists and price alerts
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WatchlistItem } from './watchlist-item';


export interface Watchlist { 
    id: number;
    name: string;
    position: number;
    items: Array<WatchlistItem>;
}